S3/requirements.txt
click==8.0.3
matplotlib==3.4.3
numpy==1.21.2
Pillow==9.0.0
python-dotenv==0.19.2
scikit_learn==1.0.2
seaborn==0.11.2
setuptools==58.0.4
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3